app/pyproject.toml
[tool.poetry]
name = "final-project"
version = "0.1.0"
description = "FastAPI for final-project"
authors = ["sonyak <kch9141@naver.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
fastapi = "0.70.0"
pandas = "^1.4.2"
uvicorn = "^0.17.6"
pydantic = "^1.9.1"
python-multipart = "^0.0.5"
Jinja2 = "^3.1.2"
requests = "^2.27.1"
Pillow = "^9.1.1"
Levenshtein = "^0.18.1"
transformers = "^4.19.2"
torch = "^1.11.0"
sentencepiece = "^0.1.96"
plotly = "^5.8.0"
tenacity = "^8.0.1"
matplotlib = "^3.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
